split components into components + pages
also split large files
